[SPARK-48754] Further refactoring to make the code more stable and accessible
### What changes were proposed in this pull request?
Refactor the internal client-related functions and structs so that we no longer expose the GRPC protocol directly to the Spark Session and the DataFrame classes. In addition, this adds additional tests to increase the overall coverage to 60%.

### Why are the changes needed?
To be able to have proper, centralized retry logic, we need to be able to decouple the consumers from the underlying GRPC stream.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
* Existing tests.
* Added additional tests for the new client module
* Cleaned up the existing tests to have a more usable surface area
* Added some additional mocks to the client handling

Closes #33 from grundprinzip/client-refactoring.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>