Improve Package Structure
### What changes were proposed in this pull request?
This patch fixes some of the issues in the current code and the way the interfaces are used and exposed to the other packages.  In particular, now all DF functions should take a context to deal with resolving schema etc dynamically and for API consistency.

Second, all columns and expressions need to support the `column.Convertible` interface to indicate that it can be converted to Proto. Lastly, the column package adds a `HasSchema` like interface that allows to pass a DF without passing the actual interface type.

Lastly it changes the `ToPlan` method on `Column` / `Expression` / `Convertible` to `ToProto` for more clarity.

### Why are the changes needed?
Ease of use.

### Does this PR introduce _any_ user-facing change?
Slight package changes. But still pre-release.

### How was this patch tested?
Existing tests.

Closes #65 from grundprinzip/package_refactorign.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Martin Grund <martin.grund@databricks.com>