[#73] Replace usage of map[string]Convertible in withColumns with a sequential alternative
### What changes were proposed in this pull request?
The Golang map type is not meant to provide an insertion order stable iteration. This means when a user calls:

```golang
df, err = df.WithColumns(ctx, map[string]column.Convertible{
		"newCol1": functions.Lit(1),
		"newCol2": functions.Lit(2),
	})
```

There is no guarantee about the order of the columns added. However, in PySpark, this is not the case, and the order is preserved. For that reason, we need to use a different way of adding multiple columns that preserve the order. 

This patch changes the interface of the `withColumns` function so that the argument is a new type called `column.Alias` that implements the `column.Convertible` interface. This allows for the following method signature.

```golang
WithColumns(ctx context.Context, alias ...column.Alias) (DataFrame, error)
```

In Spark code, the function can be used as follows:

```golang
df, err = df.WithColumns(
    ctx, column.WithAlias("newCol1", functions.Lit(1)),
    column.WithAlias("newCol2", functions.Lit(2)))
```

Which provides a similarly convenient way of creating a sequence of new columns for the function.

### Why are the changes needed?
Correctness

### Does this PR introduce _any_ user-facing change?
Changes the signature of the `withColumn` function.

### How was this patch tested?
Added UT and fixed integration test.