(FIX Issue #87) Correct Config to accurately set UserContext
The proto lacked userContext which led to not able to set values correctly for current session & user context.
### What changes were proposed in this pull request?
client.Config() function proto lacked UserContext

### Why are the changes needed?
photo requires a user context that was actually omitted in the development and was not caught since the call seemingly work, but does not set the value in the correct context.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Setting spark.sql.ansi.enabled to true (default false) and running a ansi violation cast as well as getting the value from spark.sql using set and using Config.Get in a golang application.

Closes #88 from magpierre/Correct-spark.Config.Set-issue-.

Authored-by: Magnus Pierre <magnus.pierre@icloud.com>
Signed-off-by: Martin Grund <martin.grund@databricks.com>